%Generated by roxygen2: do not edit by hand
%Please edit documentation in R/pack7-internal.R
\docType{package}
\name{pack7}
\alias{pack7}
\alias{pack7-package}
\title{Introduction to machine learning in R}
\description{
This function it should take a formula object as well as a dataset and return a ridgereg object.
}
\author{
Jorge Villar Fuentes and Ashmaful Alam
}
