% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}

\item{lambda}{numeric}

\item{x_norm}{matrix}

\item{beta_ridge}{numeric}

\item{y_hat}{numeric}

\item{data_name}{character}
}
\value{
a ridgereg object.
}
\description{
The  function it should take a formula object as well as a dataset and return a ridgereg object.
}

