% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize airport arrival delays}
\value{
a plot
}
\description{
Creates a plot that visualizes the mean delay of ﬂights for diﬀerent airports by longitude and latitude.
}
